updated java program to have some more sanity checks and removed redundant file browser

+++ java-files/wizard.java
@@ -6,6 +6,7 @@ import java.nio.file.*;
 import java.awt.BorderLayout; 
 import java.io.IOException;
 import java.nio.file.attribute.BasicFileAttributes;
+import java.io.File.*;
 
 public class wizard extends JFrame implements ActionListener
 {
@@ -18,7 +19,6 @@ public class wizard extends JFrame implements ActionListener
 	private JMenuBar jvBar; 
 	private JLabel readMeLabel; 
 	private JPanel panel; 
-	private JFileChooser selectInstall; 
 	private Path currentPath; 
 	private Path testPath;
 
@@ -60,8 +60,7 @@ public class wizard extends JFrame implements ActionListener
 		removeMod = new JMenuItem("Remove Mod"); 
 		exitProgram = new JMenuItem("Exit"); 
 		jvBar = new JMenuBar(); 
-		selectInstall = new JFileChooser("Browse Install Location"); 
-		setDefaultOSDirectory(selectInstall); 
+		setDefaultOSDirectory(); 
 		readMeLabel = new JLabel("<html> This is the setup program for Age of Empires 3 Overhauled Edition. <br> <br> This DOES NOT work for retail copies, only Steam versions. This program WILL NOT work if you move the files from the mod's path or have a non-standard AoE3 install, you will need to do a manual install then. <br> <br> Clicking on \"Install Mod\" will create a backup of your Age Of Empires 3 data files in \"./Age of Empires 3/bin/data\" in the \"./Age of Empires 3/bin/databak\" folder. <br> <br> Clicking on \"Remove Mod\" will override your data files with the backup created earlier. <br> <br> Default Installation Paths. WIN: C:\\Program Files (x86)\\Steam\\Steamapps\\common\\Age Of Empires 3\\ LINUX: ~/.local/.share/Steam/steamapps/common/Age Of Empires 3/ </html>");
 		panel = new JPanel();
 }
@@ -76,15 +75,13 @@ public class wizard extends JFrame implements ActionListener
 		if (e.getSource() == itemHelp)
 			//show the read me once the user clicks on the menu button
 			JOptionPane.showMessageDialog(null, readMeLabel, "READ ME", JOptionPane.INFORMATION_MESSAGE);
+
 		else if (e.getSource() == installMod)
 		{
 			//show the file navigator once the user clicks on the install mod button in the menu
-			int returnVal = selectInstall.showOpenDialog(wizard.this); 
 			System.out.println("This current path is:" + currentPath);
 			System.out.println("Test path is:" + testPath);
-
-			if (returnVal == JFileChooser.APPROVE_OPTION)
-			{
+			
 				//make sure that the user is in the right path to copy the files, and not somewhere else
 				if (currentPath.toString().equals(testPath.toString()) == false)
 					System.out.println("it didn't work! You're probably not in the right directory."); 
@@ -108,14 +105,16 @@ public class wizard extends JFrame implements ActionListener
 					}
 
 
-				}		
-			}
+				}
+		}
+			
 
-	}
-	else if(e.getSource() == removeMod)
-	{
-		System.out.println("This current path is:" + currentPath);
-		System.out.println("Test path is:" + testPath);
+	
+		else if(e.getSource() == removeMod)
+		{
+			System.out.println("This current path is:" + currentPath);
+			System.out.println("Test path is:" + testPath);
+			Path deleteCheckPath = (Paths.get(currentPath +"/bin/data/civs.xml"));
 			//make sure that the user is in the right path to copy the files, and not somewhere else
 			if (currentPath.toString().equals(testPath.toString()) == false)
 				System.out.println("it didn't work! You're probably in the wrong directory!"); 
@@ -124,37 +123,52 @@ public class wizard extends JFrame implements ActionListener
 				System.out.println("modpath is: " + modPath); 
 				System.out.println("backup path is: " + backupData);
 				System.out.println("data path is: "+ dataPath); 
-				try
-				{
+				if (Files.exists(deleteCheckPath) && Files.isRegularFile(deleteCheckPath))
+					try
+					{
 					//delete the modified directory and replace it with the default one
-            				deleteDirectoryJava7(dataPath.toString());
-					copyDirectoryFileVisitor(backupData.toString(), dataPath.toString());
-					JOptionPane.showMessageDialog(null, "File Restore has completed", "Message", JOptionPane.INFORMATION_MESSAGE); 
+            					deleteDirectoryJava7(dataPath.toString());
+						copyDirectoryFileVisitor(backupData.toString(), dataPath.toString());
+						JOptionPane.showMessageDialog(null, "File Restore has completed", "Message", JOptionPane.INFORMATION_MESSAGE); 
 
-				}
-				catch(IOException ex)
+					}
+					catch(IOException ex)
+					{
+						System.out.println(ex.toString()); 
+					}
+				else if(!Files.exists(deleteCheckPath))
 				{
-					System.out.println(ex.toString()); 
+					try
+					{
+						//since we're trying to delete a file that doesn't exist, we're assuming that the user doesn't have the mod installed and did a misclick. In that case, we back up existing files.
+						copyDirectoryFileVisitor(dataPath.toString(), backupData.toString());
+					}
+					catch(IOException ex)
+					{
+						System.out.println(ex.toString()); 
+
+					}
+
 				}
 
 
 			}		
-	}
-	else if(e.getSource() == exitProgram)
-		System.exit(0);
+		}
+		else if(e.getSource() == exitProgram)
+			System.exit(0);
 	
 
-}
+	
+	}
 
 
-	private void setDefaultOSDirectory(JFileChooser fileChooser)
+	private void setDefaultOSDirectory()
 	{
 		//method to see what OS the user is on. Due to AoE3 being deprecated on MacOS, it's not a supported platform. And given that the game is no longer sold in retail stores, only Steam is being currently supported for now
 		String thisOS = System.getProperty("os.name"); 
 		if (thisOS.contains("Windows") || thisOS.contains("windows"))
 		{
 			System.out.println("User's current OS is a Windows variant, setting default directory");
-			fileChooser.setCurrentDirectory(new java.io.File("C:/Program Files (x86}/Steam/Steamapps/Common/Age of Empires 3/")); 
 			currentPath = Paths.get("C:/Program Files (x86}/Steam/Steamapps/Common/Age of Empires 3/");
 			testPath = Paths.get("C:/Program Files (x86}/Steam/Steamapps/Common/Age of Empires 3/");
 		}
@@ -162,14 +176,11 @@ public class wizard extends JFrame implements ActionListener
 		{
 			System.out.println("User's current OS is a Linux variant, setting default directory"); 
 			String home = System.getProperty("user.home");
-			fileChooser.setCurrentDirectory(new java.io.File(home + "/.local/share/Steam/steamapps/common/Age Of Empires 3/"));
 			currentPath = Paths.get(home + "/.local/share/Steam/steamapps/common/Age Of Empires 3/"); 
 			testPath = Paths.get(home + "/.local/share/Steam/steamapps/common/Age Of Empires 3/"); 
 
 
 		}
-		fileChooser.setFileSelectionMode(JFileChooser.DIRECTORIES_ONLY); 
-		fileChooser.setDialogTitle("Mod Install Location"); 
 		}
